refactor(query): extract discovery engine into importable discovery/ packages
The tofu-native discovery engine (schema introspection → filtered
data-source query → deterministic import-block emission → generate-config
cleanup) lived under query/internal/{tofu,schema,query,clean,importblock}.
Go forbids importing another module's internal/ packages, so the upcoming
MCP server (its own module) could not reuse it without shelling out to the
terrapod-query binary or duplicating the logic.

Promote the five packages out of internal/ into a public discovery/
namespace so both consumers import the same code: the standalone
terrapod-query CLI (unchanged) and the future mcp/ module.

Mechanical and behaviour-preserving: git-mv the packages (tests + testdata
move with them), repoint every import path (query/internal/X →
query/discovery/X) plus the cross-package imports and the docstring
references. No API, flag, output, or licensing change; the CLI is identical.

Part of #846.

Closes #849
